# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/me": {
      "get": {
        "operationId": "UsersController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "UsersController_updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/authenticators/applications": {
      "post": {
        "operationId": "AuthenticatorsController_submitApplication",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyAuthenticatorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Authenticators"
        ]
      }
    },
...
    "/api/authenticators": {
      "get": {
        "operationId": "AuthenticatorsController_list",
        "parameters": [
          {
            "name": "category",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Authenticators"
        ]
      }
    },
    "/api/authenticators/me": {
      "patch": {
        "operationId": "AuthenticatorsController_updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAuthenticatorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Authenticators"
        ]
      }
    },
    "/api/authenticators/me/branches": {
      "get": {
        "operationId": "AuthenticatorsController_listMyBranches",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Authenticators"
        ]
      },
      "post": {
        "operationId": "AuthenticatorsController_createBranch",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBranchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Authenticators"
        ]
      }
    },
    "/api/authenticators/me/branches/{branchId}": {
      "patch": {
        "operationId": "AuthenticatorsController_updateBranch",
        "parameters": [
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBranchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Authenticators"
        ]
      },
      "delete": {
        "operationId": "AuthenticatorsController_deleteBranch",
        "parameters": [
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Authenticators"
        ]
      }
    },
    "/api/authenticators/{id}/branches": {
      "get": {
        "operationId": "AuthenticatorsController_listBranches",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Authenticators"
        ]
      }
    },
    "/api/authenticators/{id}": {
      "get": {
        "operationId": "AuthenticatorsController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Authenticators"
        ]
      }
    },
...
    "/api/banners": {
      "get": {
        "operationId": "BannersController_listPublic",
        "parameters": [
          {
            "name": "audience",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Banners"
        ]
      }
    },
    "/api/admin/banners": {
      "get": {
        "operationId": "AdminBannersController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminBanners"
        ]
      },
      "post": {
        "operationId": "AdminBannersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBannerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AdminBanners"
        ]
      }
    },
    "/api/admin/banners/{id}": {
      "patch": {
        "operationId": "AdminBannersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBannerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminBanners"
        ]
      },
      "delete": {
        "operationId": "AdminBannersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminBanners"
        ]
      }
    },
...
  },
  "info": {
    "title": "Certifine API",
    "description": "C2C marketplace with per-category third-party authentication. Generated from the running app — do not hand-edit, and do not hand-edit the client generated from it.",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateMeDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "object"
          },
          "avatarOriginalUrl": {
            "type": "object"
          },
          "avatarCropZoom": {
            "type": "object"
          },
          "avatarCropX": {
            "type": "object"
          },
          "avatarCropY": {
            "type": "object"
          },
          "interests": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "ApplyAuthenticatorDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string"
          },
          "storeName": {
            "type": "string"
          },
          "categories": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "HANDBAG",
                "IPHONE",
                "POKEMON_CARD",
                "WATCH",
                "SNEAKER",
                "DESIGNER_TOY",
                "OTHER"
              ]
            }
          },
          "yearsExperience": {
            "type": "number"
          },
          "bio": {
            "type": "string"
          },
          "feeRatePct": {
            "type": "number"
          },
          "feeMinHKD": {
            "type": "number"
          },
          "locationAddress": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "credentialDocs": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "eAndOProofDoc": {
            "type": "string"
          },
          "eAndOExpiresAt": {
            "type": "string"
          }
        },
        "required": [
          "displayName"
        ]
      },
      "UpdateAuthenticatorDto": {
        "type": "object",
        "properties": {
          "feeRatePct": {
            "type": "number"
          },
          "feeMinHKD": {
            "type": "number"
          },
          "bio": {
            "type": "string"
          },
          "yearsExperience": {
            "type": "number"
          },
          "locationAddress": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "businessHours": {
            "type": "string"
          },
          "acceptsMeetup": {
            "type": "boolean"
          }
        }
      },
      "CreateBranchDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "fullAddress": {
            "type": "string"
          },
          "districtKey": {
            "type": "string"
          },
          "businessHours": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "contactPhone": {
            "type": "string"
          },
          "contactWhatsapp": {
            "type": "string"
          },
          "isPrimary": {
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "fullAddress",
          "districtKey"
        ]
      },
      "UpdateBranchDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "fullAddress": {
            "type": "string"
          },
          "districtKey": {
            "type": "string"
          },
          "businessHours": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "contactPhone": {
            "type": "string"
          },
          "contactWhatsapp": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "isPrimary": {
            "type": "boolean"
          }
        }
      },
...
      "CreateBannerDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "severity": {
            "type": "object"
          },
          "audience": {
            "type": "object"
          },
          "isActive": {
            "type": "boolean"
          },
          "startsAt": {
            "type": "object"
          },
          "endsAt": {
            "type": "object"
          },
          "dismissible": {
            "type": "boolean"
          },
          "priority": {
            "type": "number"
          }
        },
        "required": [
          "message",
          "severity"
        ]
      },
      "UpdateBannerDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "severity": {
            "type": "object"
          },
          "audience": {
            "type": "object"
          },
          "isActive": {
            "type": "boolean"
          },
          "startsAt": {
            "type": "object"
          },
          "endsAt": {
            "type": "object"
          },
          "dismissible": {
            "type": "boolean"
          },
          "priority": {
            "type": "number"
          }
        }
      },
...
    }
  }
}